+++ public/css/sequenceserver.css
@@ -378,7 +378,7 @@ input[type=checkbox] {
   margin: 20px 0;
 }
 .resultn {
-  margin-bottom: 25px;
+  margin-top: 25px;
 }
 
 .graphical-overview {
